Reorganize UI components in docs
