+++ docs/src/components/logo.js
@@ -1,0 +1,21 @@
+import React from 'react'
+
+export default () => (
+  <div>
+    <img src="https://mdx-logo.now.sh" width="70" alt="MDX logo" />
+    <a
+      style={{
+        position: 'absolute',
+        right: 11,
+        top: 13
+      }}
+      href="https://github.com/mdx-js/mdx"
+    >
+      <img
+        src="https://icon.now.sh/github/0067ee"
+        alt="github logo"
+        width="20"
+      />
+    </a>
+  </div>
+)
