+++ docs/src/components/redirect.js
@@ -1,0 +1,4 @@
+import React from 'react'
+import {Redirect} from '@reach/router'
+
+export default props => <Redirect {...props} noThrow />
