+++ docs/src/pages/advanced/specification.md
@@ -1,3 +1,3 @@
-import { Redirect } from '../../ui'
+import Redirect from '../../components/redirect'
 
 <Redirect to='/advanced/ast' />
