+++ docs/src/pages/advanced/sync-api.md
@@ -1,3 +1,3 @@
-import { Redirect } from '../../ui'
+import Redirect from '../../components/redirect'
 
 <Redirect to='/advanced/api#sync-api' />
