+++ docs/src/pages/advanced/writing-a-plugin.md
@@ -1,3 +1,3 @@
-import { Redirect } from '../../ui'
+import Redirect from '../../components/redirect'
 
 <Redirect to="/guides/custom-loader" />
