+++ docs/src/pages/getting-started/typescript.md
@@ -1,3 +1,3 @@
-import { Redirect } from '../../ui'
+import Redirect from '../../components/redirect'
 
 <Redirect to='/advanced/typescript' />
