+++ docs/src/pages/plugins.md
@@ -1,3 +1,3 @@
-import { Redirect } from '../ui'
+import Redirect from '../components/redirect'
 
 <Redirect to='/advanced/plugins' />
